+++ octomap/src/CountingOcTree.cpp
@@ -125,7 +125,7 @@ namespace octomap {
 
     if (depth < max_depth && node->hasChildren()) {
 
-      unsigned short int center_offset_key = this->tree_max_val >> (depth + 1);
+      unsigned int center_offset_key = this->tree_max_val >> (depth + 1);
       OcTreeKey search_key;
 
       for (unsigned int i=0; i<8; ++i) {
